improve bash and create commonCheckerUtils

+++ checker/bin/spark-checker.sh
@@ -33,6 +33,7 @@ PARTITIONS
 
 -h  display this help."
 
+ALLUXIO_CHECKER_JAR=""
 PARTITIONS="${2:-10}";
 SPARK_MASTER="$1";
 SPARK_SUBMIT="";
@@ -40,17 +41,27 @@ SPARK_SUBMIT="";
 # Find the location of spark-submit in order to run the Spark job
 function find_spark_path() {
   { # Try SPARK_HOME
-    [ -f "${SPARK_HOME}/bin/spark-submit" ] && SPARK_SUBMIT="${SPARK_HOME}/bin"
+    [ -f "${SPARK_HOME}/bin/spark-submit" ] && SPARK_SUBMIT="${SPARK_HOME}/bin/spark-submit"
   } || { # Try SPARKPATH
-    [ -f "${SPARKPATH}/spark-submit" ] && SPARK_SUBMIT="${SPARKPATH}"
-  } || { # Try PATH
-    IFS=':' read -ra PATHARR <<< "$PATH"
-    for p in "${PATHARR[@]}"; do
-      if [ -f "$p/spark-submit" ]; then
-        SPARK_SUBMIT="$p"
+    [ -f "${SPARKPATH}/spark-submit" ] && SPARK_SUBMIT="${SPARKPATH}/spark-submit"
+  } || {
+    if [[ "${SPARK_SUBMIT}" == "" ]]; then
+      array=(`find "${SPARK_HOME}" -type f -name 'spark-submit'`)
+      for i in "${array[@]}"; do
+        SPARK_SUBMIT="$i"
         break;
-      fi
-    done
+      done
+    fi
+    # Try PATH
+    if [[ "${SPARK_SUBMIT}" == "" ]]; then
+      IFS=':' read -ra PATHARR <<< "$PATH"
+      for p in "${PATHARR[@]}"; do
+        if [ -f "$p/spark-submit" ]; then
+          SPARK_SUBMIT="$p/spark-submit"
+          break;
+        fi
+      done
+    fi
     if [[ "${SPARK_SUBMIT}" == "" ]]; then
       echo -e "Please set SPARK_HOME or SPARKPATH before running Spark integration checker." >&2
       exit 1
@@ -60,21 +71,38 @@ function find_spark_path() {
 
 function trigger_spark_cluster() {
   # Client mode
-  ${LAUNCHER} "${SPARK_SUBMIT}/spark-submit" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
-    --deploy-mode client "${BIN}/../target/alluxio-checker-${VERSION}-jar-with-dependencies.jar" --partitions "${PARTITIONS}"
+  ${LAUNCHER} "${SPARK_SUBMIT}" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
+    --deploy-mode client "${ALLUXIO_CHECKER_JAR}" --partitions "${PARTITIONS}"
 
   # Cluster mode
-  ${LAUNCHER} "${SPARK_SUBMIT}/spark-submit" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
-    --deploy-mode cluster "${BIN}/../target/alluxio-checker-${VERSION}-jar-with-dependencies.jar" --partitions "${PARTITIONS}"
+  ${LAUNCHER} "${SPARK_SUBMIT}" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
+    --deploy-mode cluster "${ALLUXIO_CHECKER_JAR}" --partitions "${PARTITIONS}"
 }
 
 function trigger_spark_local() {
-  ${LAUNCHER} "${SPARK_SUBMIT}/spark-submit" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
-    "${BIN}/../target/alluxio-checker-${VERSION}-jar-with-dependencies.jar" --partitions "${PARTITIONS}"
+  ${LAUNCHER} "${SPARK_SUBMIT}" --class alluxio.checker.SparkIntegrationChecker --master "${SPARK_MASTER}" \
+    "${ALLUXIO_CHECKER_JAR}" --partitions "${PARTITIONS}"
 }
 
 function main {
+  # Check if an input argument is -h
+  for i in "$@"; do
+    if [[ "$i" == "-h" ]]; then
+      echo -e "${USAGE}" >&2
+      exit 0
+    fi
+  done
+
+  # Check if too many arguments has been passed in
+  if [ "$#" -gt 2 ]; then
+    echo -e "${USAGE}" >&2
+    echo "Too many arguments passed in."
+    exit 1
+  fi
+
   source "${BIN}/../../libexec/alluxio-config.sh"
+  ALLUXIO_CHECKER_JAR="${BIN}/../target/alluxio-checker-${VERSION}-jar-with-dependencies.jar"
+
   [ -f "./SparkIntegrationReport.txt" ] && rm "./SparkIntegrationReport.txt"
   case "${SPARK_MASTER}" in
     local*)
